add all syscalls

+++ src/subsystem/resources/motion_manager/s2_move.rs
@@ -402,7 +402,7 @@ impl ScaleMotionContainer {
         if self.current_id > 0 {
             self.current_id -= 1;
         }
-        self.allocation_pool[self.current_id as usize] = self.motions[i].get_id() as u16;
+        self.allocation_pool[self.current_id as usize] = self.motions[i].get_id();
 
         Ok(())
     }
